feat: add form to create bookings to MapScreen
