+++ android/app/src/main/java/io/github/parkwithease/parkeasy/ui/search/map/MapScreen.kt
@@ -1,5 +1,6 @@
 package io.github.parkwithease.parkeasy.ui.search.map
 
+import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Refresh
@@ -13,7 +14,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.mapbox.mapboxsdk.geometry.LatLng
@@ -27,9 +31,11 @@ import io.github.parkwithease.parkeasy.R
 import io.github.parkwithease.parkeasy.model.Spot
 import io.github.parkwithease.parkeasy.ui.common.PreviewAll
 import io.github.parkwithease.parkeasy.ui.navbar.NavBar
+import io.github.parkwithease.parkeasy.ui.search.CreateBookingScreen
 import io.github.parkwithease.parkeasy.ui.search.DefaultLatitude
 import io.github.parkwithease.parkeasy.ui.search.DefaultLongitude
 import io.github.parkwithease.parkeasy.ui.search.SearchViewModel
+import io.github.parkwithease.parkeasy.ui.search.rememberCreateHandler
 import io.github.parkwithease.parkeasy.ui.search.rememberSearchHandler
 import io.github.parkwithease.parkeasy.ui.theme.ParkEasyTheme
 
@@ -42,31 +48,55 @@ fun MapScreen(
 ) {
     val loggedIn by viewModel.loggedIn.collectAsState(true)
     val latestOnNavigateToLogin by rememberUpdatedState(onNavigateToLogin)
-    val isRefreshing by viewModel.isRefreshing.collectAsState()
 
     if (!loggedIn) {
         LaunchedEffect(Unit) { latestOnNavigateToLogin() }
     } else {
-        @Suppress("unused") val handler = rememberSearchHandler(viewModel)
+        @Suppress("unused") val searchHandler = rememberSearchHandler(viewModel)
+        val createHandler = rememberCreateHandler(viewModel)
+        val cars by viewModel.cars.collectAsState()
         val spots by viewModel.spots.collectAsState()
+        val isRefreshing by viewModel.isRefreshing.collectAsState()
+
+        var showForm by rememberSaveable { mutableStateOf(false) }
+
+        BackHandler(enabled = showForm) { showForm = false }
 
         LaunchedEffect(Unit) { viewModel.onRefresh() }
-        MapScreen(
-            map = { mapSpots, mapModifier -> MapLibreMap(mapSpots, mapModifier) },
-            spots = spots,
-            isRefreshing = isRefreshing,
-            onRefresh = viewModel::onRefresh,
-            navBar = navBar,
-            snackbarHost = { SnackbarHost(hostState = viewModel.snackbarState) },
-            modifier = modifier,
-        )
+
+        if (showForm)
+            CreateBookingScreen(
+                cars = cars,
+                state = viewModel.createState,
+                handler = createHandler,
+                getSelectedIds = { viewModel.createState.selectedIds.value },
+                disabledIds = viewModel.createState.disabledIds.value,
+            )
+        else
+            MapScreen(
+                map = { mapSpots, onSpotClick, mapModifier ->
+                    MapLibreMap(mapSpots, onSpotClick, mapModifier)
+                },
+                spots = spots,
+                onSpotClick = {
+                    createHandler.reset()
+                    createHandler.onSpotChange(it)
+                    showForm = true
+                },
+                isRefreshing = isRefreshing,
+                onRefresh = viewModel::onRefresh,
+                navBar = navBar,
+                snackbarHost = { SnackbarHost(hostState = viewModel.snackbarState) },
+                modifier = modifier,
+            )
     }
 }
 
 @Composable
 fun MapScreen(
-    map: @Composable ((spots: List<Spot>, modifier: Modifier) -> Unit),
+    map: @Composable ((spots: List<Spot>, onSpotClick: (Spot) -> Unit, modifier: Modifier) -> Unit),
     spots: List<Spot>,
+    onSpotClick: (Spot) -> Unit,
     isRefreshing: Boolean,
     onRefresh: () -> Unit,
     navBar: @Composable (() -> Unit),
@@ -79,12 +109,12 @@ fun MapScreen(
         snackbarHost = snackbarHost,
         floatingActionButton = { RefreshButton(isRefreshing = isRefreshing, onRefresh = onRefresh) },
     ) { innerPadding ->
-        Surface(modifier = Modifier.padding(innerPadding)) { map(spots, modifier) }
+        Surface(modifier = Modifier.padding(innerPadding)) { map(spots, onSpotClick, modifier) }
     }
 }
 
 @Composable
-fun MapLibreMap(spots: List<Spot>, modifier: Modifier = Modifier) {
+fun MapLibreMap(spots: List<Spot>, onSpotClick: (Spot) -> Unit, modifier: Modifier = Modifier) {
     val cameraState = CameraState.Centered(latitude = DefaultLatitude, longitude = DefaultLongitude)
     val mapViewCamera = rememberSaveableMapViewCamera(MapViewCamera(state = cameraState))
 
@@ -97,6 +127,7 @@ fun MapLibreMap(spots: List<Spot>, modifier: Modifier = Modifier) {
             Symbol(
                 center = LatLng(it.location.latitude, it.location.longitude),
                 imageId = R.drawable.location,
+                onTap = { onSpotClick(it) },
             )
         }
     }
@@ -121,8 +152,9 @@ fun RefreshButton(isRefreshing: Boolean, onRefresh: () -> Unit, modifier: Modifi
 private fun MapScreenPreview() {
     ParkEasyTheme {
         MapScreen(
-            map = { _, _ -> },
+            map = { _, _, _ -> },
             spots = emptyList<Spot>(),
+            onSpotClick = {},
             isRefreshing = false,
             onRefresh = {},
             navBar = { NavBar() },
